queryInterface module becomes sparqly
