+++ src/app/query/query.html
@@ -1,4 +1,4 @@
-<div id="query-interface" class="row" data-ng-controller="QueryController">
+<div id="query-interface" class="row">
 	<div class="col-xs-12">
 		<form novalidate>
 			<!-- ux assisted form -->
