+++ src/app/query/sparqly.config.js
@@ -1,0 +1,15 @@
+//////// Module Definition ////////////
+define([
+	'query/sparqly.module',
+	'query/queryIS.service',
+	'query/query.controller'
+], function() {
+///////////////////////////////////////
+
+angular.module('sparqly').config(['$routeProvider', function($routeProvider) {
+	$routeProvider.when('/query', {
+		'templateUrl': 'app/query/query.html',
+		'controller': 'QueryController'
+	}); 
+}]); 
+return {};}); // module definition end
