+++ src/app/query/sparqly.module.js
@@ -1,0 +1,14 @@
+//////// Module Definition ////////////
+define([
+	'angular',
+], function() {
+///////////////////////////////////////
+
+
+
+angular.module('sparqly', [
+
+
+]);
+
+return {};}); // module definition end
